move some methods from group_dao to groups_scanner
