+++ google/cloud/security/common/data_access/group_dao.py
@@ -16,8 +16,6 @@
 
 from Queue import Queue
 
-import anytree
-
 from google.cloud.security.common.data_access import dao
 from google.cloud.security.common.data_access.sql_queries import select_data
 from google.cloud.security.common.util import log_util
@@ -106,74 +104,3 @@ class GroupDao(dao.Dao):
                 if member.get('member_type') == 'GROUP':
                     queue.put(member.get('member_id'))
         return all_members
-
-    def get_recursive_members(self, starting_node, timestamp):
-        """Get all the recursive members of a group.
-
-        Args:
-            starting_node: Member node from which to start getting the recursive
-                members.
-            timestamp: String of snapshot timestamp, formatted as
-                YYYYMMDDTHHMMSSZ.
-
-        Returns:
-            starting_node: Member node with all its recursive members.
-        """
-        queue = Queue()
-        queue.put(starting_node)
-
-        while not queue.empty():
-            queued_node = queue.get()
-            members = self.get_group_members('group_members',
-                                             queued_node.member_id,
-                                             timestamp)
-
-            for member in members:
-                member_node = MemberNode(member.get('member_id'),
-                                         member.get('member_email'),
-                                         member.get('member_type'),
-                                         member.get('member_status'),
-                                         queued_node)
-                if member_node.member_type == 'GROUP':
-                    queue.put(member_node)
-
-        return starting_node
-
-    def build_group_tree(self, timestamp):
-        """Build a tree of all the groups in the organization.
-
-        Args:
-            timestamp: String of snapshot timestamp, formatted as
-                YYYYMMDDTHHMMSSZ.
-
-        Returns:
-            The root node that holds the tree structure of all the groups
-                in the organization.
-        """
-        root = MemberNode(MY_CUSTOMER, MY_CUSTOMER)
-
-        all_groups = self.get_all_groups('groups', timestamp)
-        for group in all_groups:
-            group_node = MemberNode(group.get('group_id'),
-                                    group.get('group_email'),
-                                    'group',
-                                    'ACTIVE',
-                                    root)
-            group_node = self.get_recursive_members(group_node, timestamp)
-
-        LOGGER.info(anytree.RenderTree(
-            root, style=anytree.AsciiStyle()).by_attr('member_email'))
-        return root
-
-
-class MemberNode(anytree.node.NodeMixin):
-    """A custom anytree node with Group Member attributes."""
-
-    def __init__(self, member_id, member_email,
-                 member_type=None, member_status=None, parent=None):
-        self.member_id = member_id
-        self.member_email = member_email
-        self.member_type = member_type
-        self.member_status = member_status
-        self.parent = parent
-        self.rules = []
